requirements.txt
pdf2image
gTTS
python-pptx
TTS